models/mask-object-sam21-small-onnx/requirements.txt
numpy<2.0
onnx
onnxruntime
pillow
